requirements.txt
opencv-python
numpy
pyserial
pygame